Move the MultiSpanProcessor into the export package, with its friends.
